requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
implicit>=0.7.1
pytest>=7.4.0
pyarrow>=12.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
pydantic>=2.0.0
